// apps/frontend/src/shared/api/generated/model/todoDTO.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Clarte API Gateway
 * Gateway for microservices
 * OpenAPI spec version: 1.0
 */

export interface TodoDTO {
  id: string;
  userId: string;
  title: string;
  description: string;
  isCompleted: boolean;
  dueDate: string;
  createdAt: string;
  updatedAt: string;
}